+++ packages/axios-extension/test/abap/ui5-abap-repository-service.test.ts
@@ -123,11 +123,15 @@ describe('Ui5AbapRepositoryService', () => {
                 .reply(200);
             const response = await destinationService.deploy({
                 archive,
+<<<<<<< HEAD
                 bsp: { name: notExistingApp },
                 showAddInfo: true
+=======
+                bsp: { name: notExistingApp }
+>>>>>>> 0782ce092f176d9633cc6eea977209d266ba2cbf
             });
             expect(response.data).toBeDefined();
-            expect(loggerMock.info).toHaveBeenCalledTimes(8); // Ensures the logFullURL method is called to support destinations
+            expect(loggerMock.info).toHaveBeenCalledTimes(7); // Ensures the logFullURL method is called to support destinations
             expect(loggerMock.warn).toHaveBeenCalledTimes(0);
             expect(loggerMock.error).toHaveBeenCalledTimes(0);
         });