+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @sap-ux/adp-tooling
 
+## 0.8.5
+
+### Patch Changes
+
+-   Updated dependencies [2e0b1a6d]
+    -   @sap-ux/logger@0.5.0
+    -   @sap-ux/axios-extension@1.10.1
+    -   @sap-ux/system-access@0.3.11
+
+## 0.8.4
+
+### Patch Changes
+
+-   2bedc697: chore(deps): update dependency @types/adm-zip to v0.5.5
+
+## 0.8.3
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.3.10
+
+## 0.8.2
+
+### Patch Changes
+
+-   Updated dependencies [ecd5275d]
+    -   @sap-ux/axios-extension@1.10.0
+    -   @sap-ux/system-access@0.3.9
+
+## 0.8.1
+
+### Patch Changes
+
+-   Updated dependencies [de8a4878]
+    -   @sap-ux/axios-extension@1.9.0
+    -   @sap-ux/system-access@0.3.8
+
 ## 0.8.0
 
 ### Minor Changes