+++ packages/control-property-editor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/control-property-editor
 
+## 0.4.13
+
+### Patch Changes
+
+-   eb0b7b37: Chore - TypeScript 5 upgrade
+
+## 0.4.12
+
+### Patch Changes
+
+-   b817c8d5: Property change is created (unsaved) on the key press but should be on blur for the adaptation projects
+
 ## 0.4.11
 
 ### Patch Changes