+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @sap-ux/telemetry
 
+## 0.2.7
+
+### Patch Changes
+
+-   Updated dependencies [76ce5c2f]
+    -   @sap-ux/project-access@1.17.2
+
+## 0.2.6
+
+### Patch Changes
+
+-   Updated dependencies [86da13e3]
+    -   @sap-ux/project-access@1.17.1
+
+## 0.2.5
+
+### Patch Changes
+
+-   Updated dependencies [2e0b1a6d]
+    -   @sap-ux/logger@0.5.0
+    -   @sap-ux/project-access@1.17.0
+    -   @sap-ux/store@0.4.2
+
+## 0.2.4
+
+### Patch Changes
+
+-   eb0b7b37: Chore - TypeScript 5 upgrade
+-   Updated dependencies [eb0b7b37]
+    -   @sap-ux/store@0.4.1
+
+## 0.2.3
+
+### Patch Changes
+
+-   Updated dependencies [de8a4878]
+    -   @sap-ux/btp-utils@0.13.0
+
 ## 0.2.2
 
 ### Patch Changes