+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-components
 
+## 1.15.4
+
+### Patch Changes
+
+-   eb0b7b37: Chore - TypeScript 5 upgrade
+
+## 1.15.3
+
+### Patch Changes
+
+-   13dcde74: Update Icons.tsx to fix typo
+
 ## 1.15.2
 
 ### Patch Changes